# api_spec.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
        "/forms/{form_id}": {
            "get": {
                "summary": "Get Form",
                "description": "Returns a form with a given form ID\n:param form_id: The ID of the form in the collection",
                "operationId": "get_form_forms__form_id__get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Form Id",
                            "type": "string"
                        },
                        "name": "form_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update Form",
                "operationId": "update_form_forms__form_id__put",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Form Id",
                            "type": "string"
                        },
                        "name": "form_id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "title": "Form Data",
                                "type": "object"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete Form",
                "operationId": "delete_form_forms__form_id__delete",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Form Id"
                        },
                        "name": "form_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/forms": {
            "get": {
                "summary": "Get All Forms",
                "description": "Returns all the forms in the forms collection",
                "operationId": "get_all_forms_forms_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Post Form",
                "description": "Creates a form using the form data supplied in the request body\n:param form_data: The form_data received in request body",
                "operationId": "post_form_forms_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "title": "Form Data",
                                "type": "object"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/forms/{form_id}/responses/{response_id}": {
            "get": {
                "summary": "Get Single Response",
                "operationId": "get_single_response_forms__form_id__responses__response_id__get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Form Id",
                            "type": "string"
                        },
                        "name": "form_id",
                        "in": "path"
                    },
                    {
                        "required": true,
                        "schema": {
                            "title": "Response Id",
                            "type": "string"
                        },
                        "name": "response_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/forms/{form_id}/responses/": {
            "post": {
                "summary": "Post Single Response",
                "description": "Add a single response for the fiven form\n:param form_id: The ID of th form in the forms collection\n:param resp_data: The response data received in the request body",
                "operationId": "post_single_response_forms__form_id__responses__post",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Form Id",
                            "type": "string"
                        },
                        "name": "form_id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "title": "Resp Data",
                                "type": "object"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/forms/{form_id}/responses": {
            "get": {
                "summary": "Get All Responses",
                "operationId": "get_all_responses_forms__form_id__responses_get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Form Id",
                            "type": "string"
                        },
                        "name": "form_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/forms/responses/sheets/{form_owner}": {
            "get": {
                "summary": "Get Sheet Urls",
                "description": "Returns the URLs of the google sheets created for the given form_owner\n:param form_owner: The email address of the creator of the form",
                "operationId": "get_sheet_urls_forms_responses_sheets__form_owner__get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Form Owner",
                            "type": "string"
                        },
                        "name": "form_owner",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "HTTPValidationError": {
                "title": "HTTPValidationError",
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
            "ValidationError": {
                "title": "ValidationError",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "loc": {
                        "title": "Location",
                        "type": "array",
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        }
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                }
            }
        }
    }
}